kadai twitter1

+++ app/Http/Controllers/TasksController.php
@@ -15,12 +15,19 @@ class TasksController extends Controller
      */
     public function index()
     {
-        //タスク一覧を、Taskモデルから全て取得
-        $tasks = Task::all();
-        // タスク一覧ビューでそれを表示、tasks.indexでtasksを$tasksにセットし直して表示するイメージ
-        return view('tasks.index',[
-            'tasks' => $tasks,
-        ]);
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+        // タスク一覧ビューでそれを表示
+        return view('tasks.index', $data,);
     }
 
     /**
@@ -52,11 +59,12 @@ class TasksController extends Controller
             'status' => 'required|max:10',   // 追加
             'content' => 'required|max:255',
               ]);
-        //
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+              
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'status' => $request->status,
+            'content' => $request->content,
+        ]);
         //リダイレクト
         return redirect('/');
     }
@@ -133,9 +141,10 @@ class TasksController extends Controller
     {
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
-        // タスクを削除
-        $task->delete();
-
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、タスクを削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
         // トップページへリダイレクトさせる
         return redirect('/');
     }
